requirements.txt
Flask==3.0.3
tensorflow==2.19.0
opencv-python-headless==4.12.0
Pillow==10.4.0
numpy==1.26.4
Werkzeug>=2.0.0
gunicorn>=20.0.0
